Temporary fix for raw strings
Before maybe switching to a new grammar, this commit fixes syntax
coloration of raw strings. There can be a flexible amount of '#'
delimiters at the start and end of a raw string. This allows arbitrary
characters in the string (except the end sequence of " and the same
amount of '#' characters as in the beginning). For more info see:
https://doc.rust-lang.org/reference/tokens.html#characters-and-strings

Signed-off-by: Max Bureck <max.bureck@fokus.fraunhofer.de>