Issue #178 Checking valid GDB in launch config + cosmetic changes
Now the Rust debug launch config checks if the specified GDB is valid.
This is done in a blocking fashion, tests show that this does not block
the UI for an unreasonable time. Making the check non-blocking would
also make the code much more complicated, since the launcher framework
does not seem to be fit for such things.

Also some cosmetic changes were done to the debug launch config. The
"Main" tab now starts with a capital letter, as all other tabs. It also
has an icon now.

Signed-off-by: Max Bureck <max.bureck@fokus.fraunhofer.de>