Issue #278 "Run test" CodeMining
Added LaunchHandler to handle RLS "rls.run" LSP commands. These
are attached to the "Run test" CodeLens message. And invoked by the
user by clicking on the CodeMining created for the CodeLens.

The LaunchHandler will try to find a cargo test launch config for the
command parameters, if no launch config is found, it a new launch config
will be created (but not saved). The launch config will then be started.