Issue #299 Defaulting to empty RLS config file path
Now in the preferences, the location to an rls.conf file will be empty.
An empty string will now skip configuration loading and create a
default initialization configuration.

This will prevent log messages about a missing config file by default,
since the RLS does not ship with an rls.config anymore. Unfortunately,
the RLS does not document at all how an
InitializeParams#initializationOptions message looks like, so we cannot
reference from the Corrosion document to an RLS page explaining the
content of an rls.conf file.