Issue #301 Preparation for eclipse+installer link.
The eclipse installer allows registering the protocol eclipse+installer
to handle installations via web links. The link must point to an Oomph
model containing an Oomph Configuration element. This configuration
information which eclipse product and which project is selected in the
Eclipse installer.
The same link can also be dropped onto the installer, if registering
the protocol is not wanted.

This commit prepares such a Configuration model.